Make fatal error on loading error of config file.
Correctness of loading the configuration file is important part of starting
process and not only. Important data for authentication process trough some
Identity Providers like google are stored in this configuration file, we need to
know this secret keys to be able to do authenctication process for users.
Saving this secret keys inside of source files is considered as security
vulnerability.